+++ tests/test_server.py
@@ -17,7 +17,7 @@ def test_all_tools_registered():
         "llm_save_session", "llm_retrospect", "llm_gain",
         "llm_set_profile", "llm_usage", "llm_savings", "llm_health", "llm_hook_health", "llm_providers",
         "llm_check_usage", "llm_update_usage", "llm_refresh_claude_usage", "llm_codex", "llm_setup",
-        "llm_cache_stats", "llm_cache_clear", "llm_quality_report",
+        "llm_cache_stats", "llm_cache_clear", "llm_quality_report", "llm_quality_guard",
         "llm_edit", "llm_rate", "llm_dashboard",
         "llm_fs_find", "llm_fs_rename", "llm_fs_edit_many", "llm_fs_analyze_context",
         "llm_team_report", "llm_team_push",